fix: split questions by type, add QueryType to BenchmarkResult, and add stratified report test
- Split 582-line questions.ts into keyword-questions.ts, relational-questions.ts,
  graph-questions.ts, and a hub re-export in questions.ts (67 lines)
- Change BenchmarkResult.query_type from string to QueryType union type
- Export aggregateByType and TypeMetrics from report.ts for testability
- Extract compareStrategies tests to benchmark.stratified.test.ts with new
  tests for aggregateByType and printStratifiedBreakdown
- benchmark.ai.test.ts reduced from 532 to 494 lines
